split up interactive generation
